Limpieza v0.2
