+++ app/Http/Controllers/ShirtController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-
 use File;
 use Illuminate\Support\Facades\Response;
 use App\Shirt;
@@ -80,60 +79,5 @@ class ShirtController extends Controller
     {
         $shirt->delete();
         return new ShirtResource($shirt);
-    }
-
-    public function getAllShirts()
-    {
-        $TshirtsFilepath = public_path() . '/img/shirts/tshirt/';
-        $LongsleeveFilepath = public_path() . '/img/shirts/longsleeve/';
-        $tshirts = [
-            'FFFFFF' => base64_encode(File::get($TshirtsFilepath . 'FFFFFF.png')),
-            '1B998B' => base64_encode(File::get($TshirtsFilepath . '1B998B.png')),
-            'ED217C' => base64_encode(File::get($TshirtsFilepath . 'ED217C.png')),
-            'FF9B71' => base64_encode(File::get($TshirtsFilepath . 'FF9B71.png')),
-            '55DDFF' => base64_encode(File::get($TshirtsFilepath . '55DDFF.png')),
-        ];
-        $longsleeve = array(
-            'FFFFFF' => base64_encode(File::get($LongsleeveFilepath . 'FFFFFF.png')),
-            '1B998B' => base64_encode(File::get($LongsleeveFilepath . '1B998B.png')),
-            'ED217C' => base64_encode(File::get($LongsleeveFilepath . 'ED217C.png')),
-            'FF9B71' => base64_encode(File::get($LongsleeveFilepath . 'FF9B71.png')),
-            '55DDFF' => base64_encode(File::get($LongsleeveFilepath . '55DDFF.png')),
-        );
-        return Response::json(array(
-            'tshirt' => $tshirts,
-            'longsleeve' => $longsleeve
-        ));
-    }
-
-    public function getShirtImage($type, $color)
-    {
-        $filepath = public_path() . '/img/shirts/' . $type . '/' . $color . '.png';
-        if (file_exists($filepath)) {
-            return Response::json(array(
-                'content' => base64_encode(File::get($filepath)),
-            ));
-        }
-    }
-
-    public function getAllDecorations()
-    {
-        $filepath = public_path() . '/img/decorations';
-        $count = sizeof(File::files($filepath));
-        $decorations = [];
-        for ($i = 0; $i < $count; $i++) {
-            array_push($decorations, ['id' => $i, 'content' => base64_encode(File::get($filepath . '/' . $i . '.png'))]);
-        }
-        return Response::json($decorations);
-    }
-
-    public function getDecorationImage($id)
-    {
-        $filepath = public_path() . '/img/decorations/' . $id . '.png';
-        if (file_exists($filepath)) {
-            return Response::json(array(
-                'decoration' => base64_encode(File::get($filepath)),
-            ));
-        }
     }
 }
